+++ crates/miden-objects/src/note/recipient.rs
@@ -1,7 +1,4 @@
-use alloc::vec::Vec;
 use core::fmt::Debug;
-
-use miden_crypto::Felt;
 
 use super::{
     ByteReader,
@@ -62,24 +59,6 @@ impl NoteRecipient {
     /// This is the public data required to create a note.
     pub fn digest(&self) -> Word {
         self.digest
-    }
-
-    /// Returns the recipient formatted to be used with the advice map.
-    ///
-    /// The format is `inputs_length || INPUTS_COMMITMENT || SCRIPT_ROOT || SERIAL_NUMBER`
-    ///
-    /// Where:
-    /// - inputs_length is the length of the note inputs
-    /// - INPUTS_COMMITMENT is the commitment of the note inputs
-    /// - SCRIPT_ROOT is the commitment of the note script (i.e., the script's MAST root)
-    /// - SERIAL_NUMBER is the recipient's serial number
-    pub fn format_for_advice(&self) -> Vec<Felt> {
-        let mut result = Vec::with_capacity(13);
-        result.push(self.inputs.num_values().into());
-        result.extend(self.inputs.commitment());
-        result.extend(self.script.root());
-        result.extend(self.serial_num);
-        result
     }
 }