+++ crates/miden-testing/src/tx_context/builder.rs
@@ -292,9 +292,9 @@ impl TransactionContextBuilder {
             },
         };
 
-        let tx_args = TransactionArgs::default().with_note_args(self.note_args);
+        let mut tx_args = TransactionArgs::default().with_note_args(self.note_args);
 
-        let mut tx_args = if let Some(tx_script) = self.tx_script {
+        tx_args = if let Some(tx_script) = self.tx_script {
             tx_args.with_tx_script_and_args(tx_script, self.tx_script_args)
         } else {
             tx_args